mediassist/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.13.0
numpy>=1.24.0
python-dateutil>=2.8.2
google-generativeai>=0.3.0
Pillow>=10.0.0
